refactor(sbb-teaser-hero, sbb-teaser-paid): move to `@sbb-esta/lyne-elements-experimental`
We are moving the `sbb-teaser-hero` and `sbb-teaser-paid` components
to the experimental package, as we have new requirements that will
result in breaking changes.
